# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Học Tập",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/parent/login": {
      "post": {
        "tags": [
          "parent"
        ],
        "summary": "Login",
        "operationId": "parent_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or wrong PIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many wrong PINs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/logout": {
      "post": {
        "tags": [
          "parent"
        ],
        "summary": "Logout",
        "operationId": "parent_logout",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or wrong PIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/parent/session": {
      "get": {
        "tags": [
          "parent"
        ],
        "summary": "Get Session",
        "operationId": "get_parent_session",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in or wrong PIN",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/build/status": {
      "get": {
        "tags": [
          "build"
        ],
        "summary": "Get Build Status",
        "operationId": "get_build_status",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildStatus"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/build/gate": {
      "get": {
        "tags": [
          "build"
        ],
        "summary": "Get Gate",
        "operationId": "get_build_gate",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateReport"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/build/gate/approve": {
      "post": {
        "tags": [
          "build"
        ],
        "summary": "Approve Gate",
        "operationId": "approve_build_gate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/hoctap__builder__gate__ApproveIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateReport"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Refused",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/build/gate/revoke": {
      "post": {
        "tags": [
          "build"
        ],
        "summary": "Revoke Gate",
        "operationId": "revoke_build_gate",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GateReport"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Setup required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AccuracyCheck": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "null before any counted verdict"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "correct": {
            "type": "integer",
            "title": "Correct"
          },
          "wrong": {
            "type": "integer",
            "title": "Wrong"
          },
          "stale": {
            "type": "integer",
            "title": "Stale",
            "description": "verdicts for an older hash (cần kiểm tra lại)"
          },
          "unchecked": {
            "type": "integer",
            "title": "Unchecked"
          },
          "sample_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Id"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size"
          },
          "min_sample": {
            "type": "integer",
            "title": "Min Sample"
          },
          "enough_sample": {
            "type": "boolean",
            "title": "Enough Sample",
            "description": "false: chưa đủ mẫu"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "value",
          "threshold",
          "correct",
          "wrong",
          "stale",
          "unchecked",
          "sample_id",
          "sample_size",
          "min_sample",
          "enough_sample"
        ],
        "title": "AccuracyCheck"
      },
      "BookPilotPages": {
        "properties": {
          "book_id": {
            "type": "string",
            "title": "Book Id"
          },
          "pilot_pages": {
            "type": "integer",
            "title": "Pilot Pages"
          }
        },
        "type": "object",
        "required": [
          "book_id",
          "pilot_pages"
        ],
        "title": "BookPilotPages"
      },
      "BuildStatus": {
        "properties": {
          "state": {
            "type": "string",
            "const": "idle",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "state"
        ],
        "title": "BuildStatus"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "FallbackCheck": {
        "properties": {
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value",
            "description": "null when there are no pilot Problems"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "problems": {
            "type": "integer",
            "title": "Problems"
          },
          "with_fallback": {
            "type": "integer",
            "title": "With Fallback"
          }
        },
        "type": "object",
        "required": [
          "passed",
          "value",
          "threshold",
          "problems",
          "with_fallback"
        ],
        "title": "FallbackCheck"
      },
...
      "GateApproval": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "approved_at": {
            "type": "string",
            "title": "Approved At"
          },
          "est_cost": {
            "type": "number",
            "title": "Est Cost"
          },
          "sample_id": {
            "type": "string",
            "title": "Sample Id"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "invalid_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Invalid Reasons",
            "description": "why it no longer holds (Vietnamese)"
          }
        },
        "type": "object",
        "required": [
          "id",
          "approved_at",
          "est_cost",
          "sample_id",
          "valid",
          "invalid_reasons"
        ],
        "title": "GateApproval"
      },
      "GateCost": {
        "properties": {
          "pilot_cost": {
            "type": "number",
            "title": "Pilot Cost",
            "description": "USD; unknown-cost calls at the per-call cap"
          },
          "reported_cost": {
            "type": "number",
            "title": "Reported Cost",
            "description": "USD reported by the CLI"
          },
          "unknown_cost_calls": {
            "type": "integer",
            "title": "Unknown Cost Calls"
          },
          "unknown_cost_cap_usd": {
            "type": "number",
            "title": "Unknown Cost Cap Usd"
          },
          "per_page": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Page"
          },
          "pilot_pages": {
            "type": "integer",
            "title": "Pilot Pages"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "remaining_pages": {
            "type": "integer",
            "title": "Remaining Pages"
          },
          "est_cost": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Cost",
            "description": "USD for the remaining pages, to the cent"
          }
        },
        "type": "object",
        "required": [
          "pilot_cost",
          "reported_cost",
          "unknown_cost_calls",
          "unknown_cost_cap_usd",
          "per_page",
          "pilot_pages",
          "total_pages",
          "remaining_pages",
          "est_cost"
        ],
        "title": "GateCost"
      },
      "GateReport": {
        "properties": {
          "has_pilot": {
            "type": "boolean",
            "title": "Has Pilot",
            "description": "false: chưa chạy thử"
          },
          "pilot_pages": {
            "type": "integer",
            "title": "Pilot Pages"
          },
          "pilot_problems": {
            "type": "integer",
            "title": "Pilot Problems"
          },
          "books": {
            "items": {
              "$ref": "#/components/schemas/BookPilotPages"
            },
            "type": "array",
            "title": "Books"
          },
          "thresholds": {
            "$ref": "#/components/schemas/GateThresholds"
          },
          "fallback": {
            "$ref": "#/components/schemas/FallbackCheck"
          },
          "accuracy": {
            "$ref": "#/components/schemas/AccuracyCheck"
          },
          "cost": {
            "$ref": "#/components/schemas/GateCost"
          },
          "checks_passed": {
            "type": "boolean",
            "title": "Checks Passed"
          },
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GateApproval"
              },
              {
                "type": "null"
              }
            ],
            "description": "the latest approval, unless revoked"
          },
          "approved": {
            "type": "boolean",
            "title": "Approved",
            "description": "a valid approval exists: Đã duyệt"
          }
        },
        "type": "object",
        "required": [
          "has_pilot",
          "pilot_pages",
          "pilot_problems",
          "books",
          "thresholds",
          "fallback",
          "accuracy",
          "cost",
          "checks_passed",
          "approval",
          "approved"
        ],
        "title": "GateReport"
      },
      "GateThresholds": {
        "properties": {
          "max_fallback_share": {
            "type": "number",
            "title": "Max Fallback Share"
          },
          "min_key_accuracy": {
            "type": "number",
            "title": "Min Key Accuracy"
          },
          "min_sample": {
            "type": "integer",
            "title": "Min Sample"
          }
        },
        "type": "object",
        "required": [
          "max_fallback_share",
          "min_key_accuracy",
          "min_sample"
        ],
        "title": "GateThresholds"
      },
...
      "LoginRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^[0-9]{4}$",
            "title": "Pin"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "pin"
        ],
        "title": "LoginRequest"
      },
...
      "SessionStatus": {
        "properties": {
          "authenticated": {
            "type": "boolean",
            "const": true,
            "title": "Authenticated"
          }
        },
        "type": "object",
        "required": [
          "authenticated"
        ],
        "title": "SessionStatus"
      },
...
      "hoctap__builder__gate__ApproveIn": {
        "properties": {
          "accept_cost": {
            "type": "boolean",
            "title": "Accept Cost"
          },
          "est_cost_seen": {
            "type": "number",
            "title": "Est Cost Seen"
          }
        },
        "type": "object",
        "required": [
          "accept_cost",
          "est_cost_seen"
        ],
        "title": "ApproveIn"
      },
...
    }
  }
}